Misc updates; Dashboard mobile top menu FE updates; Updates

+++ resources/views/partials/dashboard/mobile-top-menu.blade.php
@@ -25,7 +25,10 @@
         @endcan
       @endif
 
-      @if (preg_match("/shop/i", env('MODULES')))
+
+      {{-- Product options --}}
+      @if (preg_match("/product/i", env('MODULES')))
+        <li class="nav-item dropdown bg-light text-dark p-3-rm border">
         <li class="nav-item dropdown bg-light text-dark p-3-rm border">
           <a class="nav-link dropdown-toggle text-dark p-3 py-4" href="#" id="mobTopMenuProductDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
             <i class="fas fa-dice-d6 text-secondry mr-3"></i>
@@ -51,7 +54,11 @@
 
           </div>
         </li>
+      @endif
 
+
+      {{-- Shop options --}}
+      @if (preg_match("/shop/i", env('MODULES')))
         <li class="nav-item dropdown bg-light text-dark p-3-rm border">
           <a class="nav-link dropdown-toggle text-dark p-3 py-4" href="#" id="mobTopMenuShopDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
             <i class="fas fa-shopping-cart text-secondry mr-3"></i>
@@ -113,88 +120,95 @@
 
           </div>
         </li>
+      @endif
 
-        {{-- CMS options --}}
-        @if (preg_match("/cms/i", env('MODULES')))
-          <li class="nav-item dropdown bg-light text-dark p-3-rm border">
-            <a class="nav-link dropdown-toggle text-dark p-3 py-4" href="#" id="mobTopMenuCmsDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-              <i class="fas fa-edit text-secondry mr-3"></i>
-              <span class="font-weight-bold">
-              CMS
-              </span>
+
+      {{-- CMS options --}}
+      @if (preg_match("/cms/i", env('MODULES')))
+        <li class="nav-item dropdown bg-light text-dark p-3-rm border">
+          <a class="nav-link dropdown-toggle text-dark p-3 py-4" href="#" id="mobTopMenuCmsDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+            <i class="fas fa-edit text-secondry mr-3"></i>
+            <span class="font-weight-bold">
+            CMS
+            </span>
+          </a>
+          <div class="dropdown-menu p-0 border-danger-rm" aria-labelledby="mobTopMenuCmsDropdown">
+
+            <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-cms-post') }}">
+              <i class="fas fa-edit mr-3"></i>
+              Post
             </a>
-            <div class="dropdown-menu p-0 border-danger-rm" aria-labelledby="mobTopMenuCmsDropdown">
+            <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-cms-webpage') }}">
+              <i class="fas fa-file mr-3"></i>
+              Page
+            </a>
+            <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-cms-gallery') }}">
+              <i class="fas fa-images mr-3"></i>
+              Gallery
+            </a>
+            <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-cms-nav-menu') }}">
+              <i class="fas fa-list mr-3"></i>
+              Nav menu
+            </a>
+            <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-cms-theme') }}">
+              <i class="fas fa-palette mr-3"></i>
+              Theme
+            </a>
 
-              <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-cms-post') }}">
-                <i class="fas fa-edit mr-3"></i>
-                Post
-              </a>
-              <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-cms-webpage') }}">
-                <i class="fas fa-file mr-3"></i>
-                Page
-              </a>
-              <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-cms-gallery') }}">
-                <i class="fas fa-images mr-3"></i>
-                Gallery
-              </a>
-              <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-cms-nav-menu') }}">
-                <i class="fas fa-list mr-3"></i>
-                Nav menu
-              </a>
-              <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-cms-theme') }}">
-                <i class="fas fa-palette mr-3"></i>
-                Theme
-              </a>
+          </div>
+        </li>
+      @endif
 
-            </div>
-          </li>
-        @endif
 
-        {{-- Calendar options --}}
-        @if (preg_match("/school/i", env('MODULES')))
-          <li class="nav-item dropdown bg-light text-dark p-3-rm border">
-            <a class="nav-link dropdown-toggle text-dark p-3 py-4" href="#" id="mobTopMenuCalendarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-              <i class="fas fa-calendar text-secondry mr-3"></i>
-              <span class="font-weight-bold">
+      {{-- Calendar options --}}
+      @if (preg_match("/school/i", env('MODULES')))
+        <li class="nav-item dropdown bg-light text-dark p-3-rm border">
+          <a class="nav-link dropdown-toggle text-dark p-3 py-4" href="#" id="mobTopMenuCalendarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+            <i class="fas fa-calendar text-secondry mr-3"></i>
+            <span class="font-weight-bold">
+            Calendar
+            </span>
+          </a>
+          <div class="dropdown-menu p-0 border-danger-rm" aria-labelledby="mobTopMenuCalendarDropdown">
+
+            <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-school-calendar') }}">
+              <i class="fas fa-calendar mr-3"></i>
               Calendar
-              </span>
             </a>
-            <div class="dropdown-menu p-0 border-danger-rm" aria-labelledby="mobTopMenuCalendarDropdown">
 
-              <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-school-calendar') }}">
-                <i class="fas fa-calendar mr-3"></i>
-                Calendar
-              </a>
+          </div>
+        </li>
+      @endif
 
-            </div>
-          </li>
-        @endif
 
-        {{-- Team options --}}
-        @if (preg_match("/team/i", env('MODULES')))
-          <li class="nav-item dropdown bg-light text-dark p-3-rm border">
-            <a class="nav-link dropdown-toggle text-dark p-3 py-4" href="#" id="mobTopMenuTeamDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-              <i class="fas fa-calendar text-secondry mr-3"></i>
-              <span class="font-weight-bold">
+      {{-- Team options --}}
+      @if (preg_match("/team/i", env('MODULES')))
+        <li class="nav-item dropdown bg-light text-dark p-3-rm border">
+          <a class="nav-link dropdown-toggle text-dark p-3 py-4" href="#" id="mobTopMenuTeamDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+            <i class="fas fa-calendar text-secondry mr-3"></i>
+            <span class="font-weight-bold">
+            Team
+            </span>
+          </a>
+          <div class="dropdown-menu p-0 border-danger-rm" aria-labelledby="mobTopMenuTeamDropdown">
+
+            <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-team') }}">
+              <i class="fas fa-calendar mr-3"></i>
               Team
-              </span>
             </a>
-            <div class="dropdown-menu p-0 border-danger-rm" aria-labelledby="mobTopMenuTeamDropdown">
 
-              <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-team') }}">
-                <i class="fas fa-calendar mr-3"></i>
-                Team
-              </a>
+            <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-quick-contacts') }}">
+              <i class="fas fa-calendar mr-3"></i>
+              Quick contacts
+            </a>
 
-              <a class="dropdown-item py-3 pl-5 font-weight-bold" href="{{ route('dashboard-quick-contacts') }}">
-                <i class="fas fa-calendar mr-3"></i>
-                Quick contacts
-              </a>
+          </div>
+        </li>
+      @endif
 
-            </div>
-          </li>
-        @endif
 
+      {{-- Report options --}}
+      @if (true)
         <li class="nav-item dropdown bg-light text-dark p-3-rm border">
           <a class="nav-link dropdown-toggle text-dark p-3 py-4" href="#" id="mobTopMenuReportDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
             <i class="fas fa-dice-d6 text-secondry mr-3"></i>
@@ -305,6 +319,8 @@
         </a>
       </li>
 
+
+      {{-- Logout and other user related options --}}
       @guest
       @else
         <li class="nav-item dropdown bg-light text-dark p-3 border">
